refactor(m12): extract CloudImageEditorState type out of SharedState
